[Refactor] rename : PurchasedLotto -> Lotto
