+++ src/main/java/lotto/model/exception/InvalidLottoNumberException.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.exception;
 
 public class InvalidLottoNumberException extends RuntimeException {
         public InvalidLottoNumberException() {
