+++ src/main/java/lotto/model/exception/LottoNumberDuplicationException.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.exception;
 
 public class LottoNumberDuplicationException extends RuntimeException {
         public LottoNumberDuplicationException() {
