+++ src/main/java/lotto/model/exception/WinnerNumbersInputFormException.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.exception;
 
 public class WinnerNumbersInputFormException extends RuntimeException {
         public WinnerNumbersInputFormException() {
