+++ src/test/java/lotto/model/PaymentTest.java
@@ -1,5 +1,7 @@
 package lotto.model;
 
+import lotto.model.exception.PaymentNotMultipleOfThousandException;
+import lotto.model.exception.PaymentNotNaturalNumberException;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
